+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:tabular',
-    version: '1.7.2-rc.0',
+    version: '1.8.0-rc.0',
     summary: 'Extends aldeed:tabular package for our applications',
     git: 'https://github.com/trychlos/pwix-tabular.git',
     documentation: 'README.md'
@@ -36,11 +36,18 @@ function configure( api ){
     _use( 'ecmascript' );
     _use( 'less@4.0.0', 'client' );
     _use( 'pwix:bootbox@1.5.5' );
+    _use( 'pwix:cookie-manager@2.2.0' );
+    _use( 'pwix:field@1.5.0' );
+    _use( 'pwix:forms@1.5.0' );
     _use( 'pwix:i18n@1.5.0' );
     _use( 'pwix:logger@1.0.0-rc' );
+    _use( 'pwix:modal@2.0.0' );
     _use( 'pwix:modal-info@1.4.0' );
+    _use( 'pwix:tabbed@1.7.0' );
+    _use( 'pwix:tolert@1.5.0' );
     _use( 'pwix:ui-bootstrap5@2.0.0' );
     _use( 'pwix:ui-fontawesome6@1.0.0', 'client' );
+    _use( 'pwix:ui-utils@1.5.0' );
     _use( 'reactive-var' );
     _use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     _use( 'tracker' );