+++ src/common/i18n/fr.js
@@ -6,6 +6,10 @@ Tabular.i18n = {
     ...Tabular.i18n,
     ...{
         fr: {
+            cookies: {
+                columns_description: 'Enregistre la liste des colonnes choisies pour être affichées dans la liste tabulaire nommée.',
+                rows_description: 'Enregistre le nombre de lignes à afficher pour chaque page de la liste tabulaire nommée.'
+            },
             delete: {
                 btn_title: 'Supprime l\'objet "%s"',
                 confirm_content: 'Vous êtes sur le point de supprimer l\'object "%s".<br />Etes-vous sûr ?',
@@ -25,7 +29,10 @@ Tabular.i18n = {
             },
             settings: {
                 columns_selection: 'Sélectionner les colonnes',
-                title: 'Personaliser'
+                dialog_title: 'Propriétés de la table',
+                recording_error: 'Vos préférences n\'ont pas pu être enregistrées. Est-il possible que vous ayez refusé le cookie ?',
+                recording_success: 'Vos préférences ont été enregistrées avec succès',
+                title: 'Personnalise les propriétés de la table'
             }
         }
     }