+++ src/common/js/constants.js
@@ -5,8 +5,9 @@
 Tabular.C = {
     // menu items
     Items: {
-        COLUMN_SELECTION: 'COLUMN_SELECTION',
-        DIVIDER: 'DIVIDER'
+        COLUMN_SELECTION:   'COLUMN_SELECTION',
+        DIVIDER:            'DIVIDER',
+        TABULAR_SETTINGS:   'TABULAR_SETTINGS'
     },
 
     // verbosity levels
@@ -25,3 +26,7 @@ Tabular.C = {
 // non exported variables
 
 I18N = 'pwix:tabular:i18n:namespace';
+
+COOKIE_RESPONSIBLE = 'pwix:tabular';
+COOKIE_COLUMNS = 'tabular-columns';
+COOKIE_ROWS_PER_PAGE = 'rows-per-page';